Moved shared data types to 'types' subcrate

+++ crates/store/src/backend/foundationdb/write.rs
@@ -72,7 +72,7 @@ impl FdbStore {
                     Operation::Collection {
                         collection: collection_,
                     } => {
-                        collection = *collection_;
+                        collection = u8::from(*collection_);
                     }
                     Operation::DocumentId {
                         document_id: document_id_,
@@ -193,7 +193,7 @@ impl FdbStore {
                     Operation::Log { collection, set } => {
                         let key = LogKey {
                             account_id,
-                            collection: *collection,
+                            collection: u8::from(*collection),
                             change_id,
                         }
                         .serialize(WITH_SUBSPACE);
