+++ crates/store/src/backend/rocksdb/write.rs
@@ -179,7 +179,7 @@ impl RocksDBTransaction<'_, '_> {
                 Operation::Collection {
                     collection: collection_,
                 } => {
-                    collection = *collection_;
+                    collection = u8::from(*collection_);
                 }
                 Operation::DocumentId {
                     document_id: document_id_,
@@ -261,7 +261,7 @@ impl RocksDBTransaction<'_, '_> {
                 Operation::Log { collection, set } => {
                     let key = LogKey {
                         account_id,
-                        collection: *collection,
+                        collection: u8::from(*collection),
                         change_id,
                     }
                     .serialize(0);
